+++ NEWS.md
@@ -13,8 +13,9 @@ output: html_document
   thereby making downstream use of the object clearer.
 * Breaking changes
     - `keep_pref_suff()` --> `get_pref_suff()` to bring consistency with `get_piece()`.
-    - `keep` --> `which` change name of argument in `get_pref_suff()`.
-* New wrapper function `get()` returns requested piece of a label.
+    - `keep` --> `which` for argument name in `get_pref_suff()`
+      to bring consistency with other functions.
+* New wrapper function `get_piece()` returns requested piece of a label.
 * Added note to README.Rmd about installing from CRAN.
 * Added project status badge.
 * Added CRAN status badge.